+++ wp-content/plugins/wpforms-lite/src/Integrations/AI/AI.php
@@ -1,18 +1,26 @@
 <?php
 
+// phpcs:disable Generic.Commenting.DocComment.MissingShort
+/** @noinspection PhpIllegalPsrClassPathInspection */
+/** @noinspection AutoloadingIssuesInspection */
+// phpcs:enable Generic.Commenting.DocComment.MissingShort
+
 namespace WPForms\Integrations\AI;
 
 use WPForms\Integrations\IntegrationInterface;
+use WPForms\Integrations\AI\Admin\Ajax\Choices as ChoicesAjax;
+use WPForms\Integrations\AI\Admin\Ajax\Forms as FormsAjax;
 use WPForms\Integrations\AI\Admin\Builder\Enqueues;
 use WPForms\Integrations\AI\Admin\Builder\FieldOption;
-use WPForms\Integrations\AI\Admin\Ajax\Choices;
+use WPForms\Integrations\AI\Admin\Builder\Forms as FormsEnqueues;
+use WPForms\Integrations\AI\Admin\Pages\Templates as TemplatesPage;
 
 /**
  * Integration of the AI features.
  *
  * @since 1.9.1
  */
-final class AI implements IntegrationInterface {
+class AI implements IntegrationInterface {
 
 	/**
 	 * Determine whether the integration is allowed to load.
@@ -23,7 +31,7 @@ final class AI implements IntegrationInterface {
 	 */
 	public function allow_load(): bool {
 
-		// Always load the Settings class in order to register the toggle.
+		// Always load the Settings class to register the toggle.
 		if ( wpforms_is_admin_page( 'settings', 'misc' ) ) {
 			( new Admin\Settings() )->init();
 		}
@@ -41,9 +49,16 @@ final class AI implements IntegrationInterface {
 		if ( wpforms_is_admin_page( 'builder' ) ) {
 			( new Enqueues() )->init();
 			( new FieldOption() )->init();
+			( new FormsEnqueues() )->init();
 		}
 
-		$this->load_ajax_classes();
+		if ( wpforms_is_admin_page( 'templates' ) ) {
+			( new TemplatesPage() )->init();
+		}
+
+		if ( wpforms_is_admin_ajax() ) {
+			$this->load_ajax_classes();
+		}
 	}
 
 	/**
@@ -51,13 +66,10 @@ final class AI implements IntegrationInterface {
 	 *
 	 * @since 1.9.1
 	 */
-	private function load_ajax_classes() {
-
-		if ( ! wpforms_is_admin_ajax() ) {
-			return;
-		}
+	protected function load_ajax_classes() {
 
 		( new FieldOption() )->init();
-		( new Choices() )->init();
+		( new ChoicesAjax() )->init();
+		( new FormsAjax() )->init();
 	}
 }